Port of https://github.com/elastic/apm-server/pull/9760

Write to a byte buffer instead of writing to the gzip writer, then compress the payload once we are about to flush it to ES. Benchmarks shows a significant improvement in cpu time and switching to an optimized gzip library improves the situation even further.

NOTE: the NoCompression benchmark shows no improvements as this change only affect the bulk indexer with compression enabled.

Updated benchmarks 1:

```
name                                   old time/op    new time/op     delta
Appender/NoCompression-20                 460ns ± 1%      452ns ± 4%    -1.79%  (p=0.028 n=8+10)
Appender/NoCompressionScaling-20          819ns ±11%      827ns ±10%      ~     (p=0.529 n=10+10)
Appender/BestSpeed-20                    1.22µs ±14%     0.48µs ± 6%   -61.02%  (p=0.000 n=10+10)
Appender/BestSpeedScaling-20             1.30µs ±12%     0.84µs ±13%   -35.15%  (p=0.000 n=10+10)
Appender/DefaultCompression-20           1.70µs ± 3%     0.49µs ± 4%   -71.26%  (p=0.000 n=10+10)
Appender/DefaultCompressionScaling-20    1.69µs ± 2%     0.85µs ±11%   -49.40%  (p=0.000 n=9+10)
Appender/BestCompression-20              1.68µs ± 3%     0.51µs ± 3%   -69.66%  (p=0.000 n=10+10)
Appender/BestCompressionScaling-20       1.69µs ± 1%     0.86µs ±13%   -49.15%  (p=0.000 n=10+10)

name                                   old speed      new speed       delta
Appender/NoCompression-20               291MB/s ± 1%    297MB/s ± 4%    +1.84%  (p=0.034 n=8+10)
Appender/NoCompressionScaling-20        165MB/s ±11%    163MB/s ± 9%      ~     (p=0.529 n=10+10)
Appender/BestSpeed-20                   108MB/s ± 5%    282MB/s ± 5%  +160.33%  (p=0.000 n=9+10)
Appender/BestSpeedScaling-20            104MB/s ±13%    159MB/s ±12%   +54.03%  (p=0.000 n=10+10)
Appender/DefaultCompression-20         79.0MB/s ± 3%  275.1MB/s ± 4%  +248.07%  (p=0.000 n=10+10)
Appender/DefaultCompressionScaling-20  79.4MB/s ± 2%  157.4MB/s ±10%   +98.17%  (p=0.000 n=9+10)
Appender/BestCompression-20            79.5MB/s ± 2%  262.3MB/s ± 3%  +229.72%  (p=0.000 n=10+10)
Appender/BestCompressionScaling-20     79.4MB/s ± 1%  156.9MB/s ±12%   +97.63%  (p=0.000 n=10+10)

name                                   old alloc/op   new alloc/op    delta
Appender/NoCompression-20                  514B ± 0%       510B ± 0%    -0.74%  (p=0.000 n=10+8)
Appender/NoCompressionScaling-20           528B ± 0%       527B ± 1%      ~     (p=0.245 n=9+10)
Appender/BestSpeed-20                      981B ± 0%       694B ± 0%   -29.28%  (p=0.000 n=9+10)
Appender/BestSpeedScaling-20               986B ± 0%       713B ± 1%   -27.71%  (p=0.000 n=9+10)
Appender/DefaultCompression-20             987B ± 1%       695B ± 0%   -29.52%  (p=0.000 n=9+10)
Appender/DefaultCompressionScaling-20      968B ± 5%       714B ± 1%   -26.24%  (p=0.000 n=10+10)
Appender/BestCompression-20                990B ± 1%       697B ± 0%   -29.61%  (p=0.000 n=10+10)
Appender/BestCompressionScaling-20         985B ± 1%       716B ± 1%   -27.30%  (p=0.000 n=8+10)

name                                   old allocs/op  new allocs/op   delta
Appender/NoCompression-20                  7.00 ± 0%       6.00 ± 0%   -14.29%  (p=0.000 n=10+10)
Appender/NoCompressionScaling-20           7.00 ± 0%       6.00 ± 0%   -14.29%  (p=0.000 n=10+10)
Appender/BestSpeed-20                      10.0 ± 0%        6.0 ± 0%   -40.00%  (p=0.000 n=10+10)
Appender/BestSpeedScaling-20               10.0 ± 0%        6.0 ± 0%   -40.00%  (p=0.000 n=10+10)
Appender/DefaultCompression-20             10.0 ± 0%        6.0 ± 0%   -40.00%  (p=0.000 n=10+10)
Appender/DefaultCompressionScaling-20      10.0 ± 0%        6.0 ± 0%   -40.00%  (p=0.000 n=10+10)
Appender/BestCompression-20                10.0 ± 0%        6.0 ± 0%   -40.00%  (p=0.000 n=10+10)
Appender/BestCompressionScaling-20         10.0 ± 0%        6.0 ± 0%   -40.00%  (p=0.000 n=10+10)
```