@marclop 2022-12-14
> Yep, I run rally and as usual I got inconsistent results. I run the benchmark twice.

So in summary, it seems that the throughput is roughly the same from the rally results. However, Rally won't give us the throughput from the APM Agent PoV, did you run any of the `apmbench` benchmarks? That may be more indicative of the impact of this change.

> Yes, sorry, I mentioned it during the weekly but forgot to report it here. This is a change in behaviour, the "flush bytes" setting will refer to the uncompressed payload size. We should update the default value to account for that and add a changelog to make sure users are aware that they might need to increase the flush bytes (if they provided a custom one) to offset the increase in network requests.

Makes sense, it would be good to see the `apmbench` results before / and with this change with 1MB `FlushBytes` and perhaps 2MB or 3MB too.