@marclop 2023-03-20
@kruskall should we close this PR? It's been open for a while and if I recall correctly, you mentioned that the approach wasn't valid anymore? Could you update this PR and / or close it?