@kruskall 2023-03-23
Let's close it for now! I have another approach that would solve the problem but I think it's better to open an issue, share the details and then schedule time to work on this.